--- source/models/currency.hpp
@@ -1,15 +1,0 @@
-#include <utility>
-
-#include<string>
-
-using namespace std;
-
-class Currency {
-public:
-    string currency;
-    double value;
-
-    Currency() {};
-
-    Currency(string cur, float val) : currency{std::move(cur)}, value{val} {}
-};
