+++ src/foundation/include/basic_controller.hpp
@@ -23,7 +23,7 @@ namespace cfx {
         pplx::task<void> shutdown();
 
         virtual void initRestOpHandlers() { 
-            /* had to be implemented by the child class */ 
+
         }
 
         std::vector<utility::string_t> requestPath(const http_request & message);
