+++ src/foundation/include/runtime_utils.hpp
