+++ src/microsvc_controller.hpp
@@ -1,13 +1,20 @@
 #pragma once
 
 #include <basic_controller.hpp>
+#include "models/currency.hpp"
 
 using namespace cfx;
 
 class MicroserviceController : public BasicController, Controller {
+private:
+    map<string, Currency *> cache;
+
 public:
-    MicroserviceController() : BasicController() {}
+    MicroserviceController(map<string, Currency *> cache) : cache{cache} {}
+
     ~MicroserviceController() {}
+
     void handleGet(http_request request) override;
+
     void initRestOpHandlers() override;
 };
